Split user service base class to basic and federated auth.
