+++ Source/AuthenticationServer.Plugins.Infrastructure/UserServiceBase.cs
@@ -1,26 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using Affecto.AuthenticationServer.Plugins.Infrastructure.Configuration;
 using IdentityServer3.Core.Models;
 
 namespace Affecto.AuthenticationServer.Plugins.Infrastructure
 {
     public abstract class UserServiceBase : IdentityServer3.Core.Services.Default.UserServiceBase
     {
-        private readonly Lazy<IFederatedAuthenticationConfiguration> federatedAuthenticationConfiguration;
-
-        protected UserServiceBase(Lazy<IFederatedAuthenticationConfiguration> federatedAuthenticationConfiguration)
-        {
-            if (federatedAuthenticationConfiguration == null)
-            {
-                throw new ArgumentNullException(nameof(federatedAuthenticationConfiguration));
-            }
-            this.federatedAuthenticationConfiguration = federatedAuthenticationConfiguration;
-        }
-
         /// <summary>
         /// This method gets called for local authentication (whenever the user uses the username and password dialog).
         /// </summary>
@@ -48,43 +34,23 @@ namespace Affecto.AuthenticationServer.Plugins.Infrastructure
         }
 
         /// <summary>
-        /// This method gets called when the user uses an external identity provider to authenticate.
-        ///             The user's identity from the external provider is passed via the `externalUser` parameter which contains the
-        ///             provider identifier, the provider's identifier for the user, and the claims from the provider for the external user.
+        /// Override this method to implement user password check.
         /// </summary>
-        /// <param name="context">The context.</param>
-        /// <returns/>
-        public override Task AuthenticateExternalAsync(ExternalAuthenticationContext context)
-        {
-            Claim userAccountName = context.ExternalIdentity.Claims.SingleOrDefault(c => c.Type == federatedAuthenticationConfiguration.Value.UserAccountNameClaim);
-            Claim userDisplayName = context.ExternalIdentity.Claims.SingleOrDefault(c => c.Type == federatedAuthenticationConfiguration.Value.UserDisplayNameClaim);
-            if (userAccountName != null && userDisplayName != null)
-            {
-                IEnumerable<Claim> userGroups = Enumerable.Empty<Claim>();
-                if (!string.IsNullOrWhiteSpace(federatedAuthenticationConfiguration.Value.GroupsClaim))
-                {
-                    userGroups = context.ExternalIdentity.Claims.Where(c => c.Type == federatedAuthenticationConfiguration.Value.GroupsClaim);
-                }
-                CreateOrUpdateExternallyAuthenticatedUser(userAccountName.Value, userDisplayName.Value, userGroups.Select(c => c.Value));
-                context.AuthenticateResult = CreateAuthenticateResult(userAccountName.Value, AuthenticationTypes.Federation, context.SignInMessage.IdP);
-            }
-            else
-            {
-                context.AuthenticateResult = new AuthenticateResult("One or more required claims were missing from the IDP's message.");
-            }
-
-            return Task.FromResult(0);
-        }
-
-        protected virtual void CreateOrUpdateExternallyAuthenticatedUser(string userName, string userDisplayName, IEnumerable<string> groups)
-        {
-        }
-
+        /// <param name="userName">Received user name.</param>
+        /// <param name="password">Received user password.</param>
+        /// <returns></returns>
         protected virtual bool IsMatchingPassword(string userName, string password)
         {
             return false;
         }
 
+        /// <summary>
+        /// Override this method to create authentication result containing data about the user as claims.
+        /// </summary>
+        /// <param name="userName">Received user name.</param>
+        /// <param name="authenticationType">Authentication type used.</param>
+        /// <param name="identityProvider">Identity provider id, "idsrv" as default. Can be replaced with e.g. external identity provider id.</param>
+        /// <returns></returns>
         protected abstract AuthenticateResult CreateAuthenticateResult(string userName, string authenticationType, string identityProvider = "idsrv");
     }
 }
